Move proxy_task tests to correct place
